email-poc/app/requirements.txt
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-discoveryengine
nicegui
pymupdf
anthropic[vertex]
